pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ag2-persona"
version = "0.1.1"
description = "Persona-based agents for AG2 (AutoGen) - enabling distinct character embodiment"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Ryan Snodgrass", email = "ryan@snodgrass.dev"}
]
maintainers = [
    {name = "Ryan Snodgrass", email = "ryan@snodgrass.dev"}
]
keywords = ["ag2", "autogen", "llm", "agents", "persona", "character", "roleplay"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    "pyautogen>=0.2.0",
    "ruamel.yaml>=0.18",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.9",
    "mypy>=1.12",
    "types-aiofiles>=24.1.0",
    "pre-commit>=4.0.1",
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/rsnodgrass/ag2-persona"
Documentation = "https://rsnodgrass.github.io/ag2-persona/"
Repository = "https://github.com/rsnodgrass/ag2-persona.git"
Issues = "https://github.com/rsnodgrass/ag2-persona/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ag2_persona*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
ag2_persona = ["py.typed"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "PL",   # pylint
    "PERF", # performance anti-patterns
    "RUF",  # ruff-specific rules
    "SIM",  # flake8-simplify
    "PTH",  # flake8-use-pathlib
]
ignore = [
    "E501",    # line too long, handled by ruff format
    "B008",    # do not perform function calls in argument defaults
    "PLC0415", # import not at top-level (intentional lazy imports for circular deps)
    "PLR0913", # too many arguments (API design choice for PersonaAgent)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
strict = true
disallow_any_generics = true
disallow_subclassing_any = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "ag2_persona.persona_agent"
disable_error_code = ["misc"]  # ConversableAgent has type Any - external library

[[tool.mypy.overrides]]
module = "ag2_persona.persona_builder"
warn_unreachable = false

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=ag2_persona",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=90",
]

[tool.coverage.run]
source = ["ag2_persona"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
